volatility/requirements.txt
pandas
boto3
psycopg2-binary
numpy